refactor(v5): move argv construction out of the god class (#878, 2/4)
Second extraction. BaseTerraformCommandHandler goes from 2225 to 1990 lines by
handing argument construction to a collaborator that does nothing else.

src/argument-builder.ts now owns two things that were tangled together in the
handler's preamble. The pure half is the commandOptions tokenizer that mirrors
task-lib's _argStringToArray (#875) and the four readers built on it -- -out=
detection, -json detection, positional-plan-file detection, and the multi-line
input parsers for -var-file/-target/-var. The stateful half is ArgumentBuilder,
which reads task inputs to produce -replace, -parallelism and the secure var
file's -var-file token.

That second half is why this is a class and not more free functions.
secureVarFileTokens() has to record the downloaded file so it can be scrubbed
before deletion (#662), which it previously did by reaching through 'this' at
the base class's temp-file state. The dependency is now a constructor
parameter, so the one place argv building touches lifecycle is visible in the
signature instead of implied by inheritance.

replaceTokens() and secureVarFileTokens() had no callers outside
buildLeadingArgs(), so they moved wholly inside rather than leaving protected
delegates behind. The other four are called at 26 sites, all retargeted at
this.argumentBuilder. Nothing was left as a pass-through: a delegate that only
forwards would keep the method count down while leaving the class exactly as
coupled as before, which is the opposite of the point.

The module-level helpers stay re-exported from the handler because Tests/ and
generic-terraform-command-handler.ts import them from there. Preserving the
entry point keeps this behaviour-only; moving importers is churn that would
make the diff harder to verify as a no-op.

Two things the pre-flight caught. RESOURCE_ADDRESS_RE and splitCommandOptions
became orphaned imports once their only users left -- the exact shape that
failed lint on #920, where a name survived in comments and defeated the
comment-skipping guard. And the first attempt at deleting the six methods was
inverted and duplicated them; the replacement script asserted 'exactly 2 copies
of each of the 6, no unrelated members in the span' before deleting, so a
144-line deletion of the wrong region failed as a precondition instead of
silently landing.

Gate: 9/9 repo gates exit 0, lint 0, 786 passing / 0 failing, per-file floor
met. argument-builder.js 91.17 stmts / 89.53 branch / 93.33 func / 93.6 lines
against a floor of 60; the handler moved 91.7 -> 92.42 lines. No test changed:
the moved code keeps its existing coverage from CommandArgsL0, L0 and
ResourceAddressL0, which is the evidence this is a pure move.
