+++ CLAUDE.md
@@ -154,6 +154,7 @@ azure-pipelines-terraform/
 | `oci-terraform-command-handler.ts`     | OCI-specific backend (HTTP backend via PAR URL) and provider auth                                                                                                                                                                                                                                                                            |
 | `environment-variables.ts`             | Helper for setting environment variables with tracking and cleanup                                                                                                                                                                                                                                                                           |
 | `temp-file-manager.ts`                 | Owns the command handler's temp-file lifecycle: the two tracking tiers (normal-completion vs cancellation-only, #650), scrub-then-unlink of each tracked path (#595), and the secure var file's scrub-before-delete (#662)                                                                                                                   |
+| `argument-builder.ts`                  | Turns task inputs into Terraform argv: the `commandOptions` tokenizer that mirrors task-lib's `_argStringToArray` (#875) and the readers built on it (`-out=`, `-json`, positional plan file), the multi-line input parsers (`-var-file`, `-target`, `-var`), and `ArgumentBuilder` for the input-reading flags (`-replace`, `-parallelism`, secure var file) |
 | `secure-file-loader.ts`                | Downloads secure var files from ADO Secure Files library                                                                                                                                                                                                                                                                                     |
 | `id-token-generator.ts`                | Generates OIDC ID tokens for Workload Identity Federation fallback                                                                                                                                                                                                                                                                           |
 | `secure-temp.ts`                       | Secure temp-file primitives: owner-only 0600 + O_EXCL on Unix, a restrictive icacls DACL on Windows (both fail closed), plus symlink-guarded `scrubFile()` zero-overwrite-before-unlink (#595) — canonical source; byte-identical copy also in TerraformDriftReportV1 and TerraformPolicyCheckV1, gated by `scripts/check-shared-modules.js` |
